refactor uav verification into a task
