+++ package.json
@@ -44,7 +44,7 @@
     "compile": "yarn hardhat compile",
     "deploy": "yarn hardhat run ./scripts/deploy-uav.js --network mainnet",
     "deploy-test": "yarn hardhat run ./scripts/deploy-uav.js --network hardhat",
-    "verify-uav": "yarn hardhat run ./scripts/verify-uav.js --network mainnet",
+    "verify-uav": "yarn hardhat verify-proposed-uav --network mainnet",
     "verify": "yarn hardhat verify --network mainnet --constructor-args ./configuration/parameters.js",
     "verify-manual": "yarn hardhat manual-verify",
     "transfer": "yarn hardhat transfer-ownership --network mainnet --uav",
